+++ packages/commonwealth/client/scripts/views/pages/discussions/DiscussionsFeedDiscovery.tsx
@@ -20,7 +20,7 @@ export const DiscussionsFeedDiscovery = (
         rel="alternate"
         type="application/atom+xml"
         title="ATOM"
-        href={`/api/feed?bulk=true&page=1&limit=20&community_id=${community}&includePinnedThreads=${includePinnedThreads}&orderBy=${orderBy}`}
+        href={`/api/feed?cursor=1&limit=20&community_id=${community}&includePinnedThreads=${includePinnedThreads}&order_by=${orderBy}`}
       />
     </Helmet>
   );
