+++ packages/commonwealth/server/routing/router.ts
@@ -18,7 +18,7 @@ import {
 } from '@hicommonwealth/adapters';
 import { AppError, query } from '@hicommonwealth/core';
 import { Community, generateTokenIdea, User } from '@hicommonwealth/model';
-import { get_feed_router } from 'server/api/get-feed-router';
+import { get_atom_feed_router } from 'server/api/get-atom-feed-router';
 import generateImageHandler from '../routes/generateImage';
 import getUploadSignature from '../routes/getUploadSignature';
 import logout from '../routes/logout';
@@ -44,9 +44,8 @@ function setupRouter(app: Express, cacheDecorator: CacheDecorator) {
   // MCP endpoint
   app.use(api.mcp.PATH, api.mcp.router);
 
-  // TODO: review and refactor to api/internal/external if necessary
-  // TODO: these routers should be decomposed into smaller routes to individual queries
-  registerRoute(router, 'get', '/feed', get_feed_router);
+  registerRoute(router, 'get', '/feed', get_atom_feed_router);
+
   registerRoute(
     router,
     'get',
